pyproject.toml
[project]
name = "aquifer"
version = "0.1.0"
description = "HIPAA de-identification engine with portable medical data format"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
readme = "README.md"
keywords = ["hipaa", "phi", "de-identification", "medical", "dental", "healthcare"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "click>=8.1",
    "PyMuPDF>=1.23",
    "python-docx>=1.1",
    "Pillow>=10.0",
    "zstandard>=0.22",
    "cryptography>=42.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "pydantic>=2.5",
]

[project.optional-dependencies]
ner = ["spacy>=3.7"]
ocr = ["pytesseract>=0.3"]
all = ["spacy>=3.7", "pytesseract>=0.3"]
dev = ["pytest>=8.0", "pytest-asyncio", "httpx", "spacy>=3.7"]

[project.scripts]
aquifer = "aquifer.cli:cli"

[project.urls]
Documentation = "https://github.com/aquifer-health/aquifer#readme"
Issues = "https://github.com/aquifer-health/aquifer/issues"
Source = "https://github.com/aquifer-health/aquifer"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
